test: validate canonical graph transitions
